Add "Never" as a valid value for .spec.digestReflectionPolicy
This way we circumvent issues with server-side apply so that users can
explicitly change this field instead of having to remove it. The
latter case might lead to the API server not removing it if another
field manager is registered for that field, causing an unintended
drift.

This commit also aligns the v1beta3 API with the latest changes done
in v1beta2.

Signed-off-by: Max Jonas Werner <mail@makk.es>